% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{AEI}
\alias{AEI}
\title{Augmented Expected Improvement}
\usage{
AEI(
  x,
  model,
  model.from = c("laGP", "km"),
  comp.model.noise.var = 0,
  beta = 0.9,
  fmin = NULL,
  x.design = NULL,
  predf = laGP::predGPsep
)
}
\arguments{
\item{predf}{}
}
\description{
Augmented Expected Improvement
}
