% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entropy search.R
\name{covMatrix}
\alias{covMatrix}
\title{Covariance matrix}
\usage{
covMatrix(X1, X2 = NULL, theta, correlation = "gauss")
}
\arguments{
\item{correlation}{}
}
\description{
Function to calculate a variance-covariance matrix for a Gaussian process
Supported correlation functions are Gaussian, matern3_2 and matern5_2
}
