% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQIoptim.R
\name{EQIoptim}
\alias{EQIoptim}
\title{EQI optimisation}
\usage{
EQIoptim(
  Xinit,
  f,
  N = 100,
  noisy = TRUE,
  sigma = 0,
  model.type = c("laGP", "km"),
  type = "UK",
  lower = 0,
  upper = 1,
  B = 2,
  beta = 0.9,
  predf = laGP::predGPsep,
  optim.multi.start = 5
)
}
\arguments{
\item{optim.multi.start}{}
}
\description{
This algorithm uses Expected Quantile Improvement (EQI) with Gaussian processes
to find the minimum of a blackbox function
}
