\name{ESMultiTask-package}
\alias{ESMultiTask-package}
\alias{ESMultiTask}
\docType{package}
\title{
\packageTitle{ESMultiTask}
}
\description{
\packageDescription{ESMultiTask}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ESMultiTask}
\packageIndices{ESMultiTask}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ESMultiTask}

Maintainer: \packageMaintainer{ESMultiTask}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
