% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eTSSO.R
\name{eTSSO}
\alias{eTSSO}
\title{eTSSO algorithm}
\usage{
eTSSO(
  Xinit,
  f,
  N = 100,
  noisy = TRUE,
  sigma = 0,
  model.type = c("laGP", "km"),
  type = "UK",
  lower = 0,
  upper = 1,
  rmin = 2,
  predf = laGP::predGPsep,
  optim.multi.start = 5
)
}
\arguments{
\item{optim.multi.start}{}
}
\description{
Extended Two-stage sequential optimisation (eTSSO) algorithm uses modified expected improvement (MEI),
as well as varying number of replicated samples to find
the minimum of a noisy blackbox function
}
