% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GP helper functions.R
\name{find_GP_optim}
\alias{find_GP_optim}
\title{Find the optimum point of a Gaussian Process, given the design}
\usage{
find_GP_optim(
  X,
  y,
  model.type = "laGP",
  correlation = "gauss",
  lower = 0,
  upper = 1,
  optimisation.type = "optim",
  optim.start = 10,
  lhs.size = 1000
)
}
\arguments{
\item{lhs.size}{}
}
\description{
Find the optimum point of a Gaussian Process, given the design
}
