% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entropy search.R
\name{GPdistance}
\alias{GPdistance}
\title{GP distance}
\usage{
GPdistance(X1, X2 = NULL, theta)
}
\arguments{
\item{theta}{}
}
\description{
Calculates the Gaussian correlation matrix between two matrices
}
