% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entropy search.R
\name{GPupdate}
\alias{GPupdate}
\title{GP update}
\usage{
GPupdate(
  X,
  X.cur,
  y.cur = NULL,
  X.add,
  y.add = NULL,
  theta,
  mean.cur,
  Sigma,
  sigma = NULL,
  mean.add = NULL,
  correlation = "gauss"
)
}
\arguments{
\item{mean.add}{}
}
\description{
Calculates mean and variance after adding in a point to a Gaussian Process
}
