% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{KQ}
\alias{KQ}
\title{Kriging Quantile}
\usage{
KQ(
  x,
  model,
  model.from = c("laGP", "km"),
  beta = 0.1,
  predf = laGP::predGPsep,
  type = "UK"
)
}
\arguments{
\item{predf}{}
}
\description{
Kriging Quantile
}
