% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{MEI}
\alias{MEI}
\title{Modified Expected Improvement}
\usage{
MEI(
  x,
  model.noisy,
  model.det = NULL,
  model.type = "laGP",
  type = "UK",
  x.design = NULL,
  y.design = NULL,
  fmin = NULL,
  predf = laGP::predGPsep
)
}
\arguments{
\item{x}{matrix of data points at which MEI should be calculated at}

\item{model.noisy}{model of class km or laGP}

\item{model.det}{model of class km or laGP}

\item{model.type}{km or laGP}

\item{type}{kriging type, e.g UK or SK}

\item{x.design}{current design matrix}

\item{y.design}{observed values from the computer model}

\item{fmin}{kriging prediction at the point with the lowest sample mean}

\item{predf}{function for making predictions from the Gaussian process}
}
\description{
Modified Expected Improvement
}
