% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQ.R
\name{MQ}
\alias{MQ}
\title{MQ algorithm}
\usage{
MQ(
  Xinit,
  f,
  N = 100,
  noisy = TRUE,
  sigma = 0,
  model.type = c("laGP", "km"),
  type = "UK",
  lower = 0,
  upper = 1,
  B = 2,
  beta = 0.1,
  predf = laGP::predGPsep,
  optim.multi.start = 5
)
}
\arguments{
\item{optim.multi.start}{}
}
\description{
MQ algorithm uses kriging quantiles with Gaussian processes to find the minimum of a blackbox function
}
