% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multitask EI.R
\name{multitask_EI}
\alias{multitask_EI}
\title{Multitask Expected Improvement}
\usage{
multitask_EI(
  Xinit,
  f,
  N = 50,
  g = NULL,
  lower = 0,
  upper = 1,
  add = "both",
  optim.multi.start = NULL
)
}
\arguments{
\item{optim.multi.start}{}
}
\description{
Multitask Expected Improvement
}
