% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OCBA.R
\name{OCBA}
\alias{OCBA}
\title{Optimal computing budget allocation algorithm}
\usage{
OCBA(
  design,
  sample.means = design$sample.mean,
  sample.sd = design$sample.sd,
  ra
)
}
\arguments{
\item{ra}{}
}
\description{
Optimal computing budget allocation algorithm
}
