% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{parallel_tempering_sample}
\alias{parallel_tempering_sample}
\title{Parallel tempering}
\usage{
parallel_tempering_sample(
  fn,
  temps.start = NULL,
  start.loc = NULL,
  n.chains = NULL,
  N = 1000,
  p = NULL,
  proposal.dist = NULL,
  walk.sd = NULL,
  bounded = FALSE,
  lower = NULL,
  upper = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Parallel tempering algorithm to sample from any given (unnormalised) PDF
}
