% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GP helper functions.R
\name{predGP_out}
\alias{predGP_out}
\title{Make predictions from a Gaussian process, under different packages}
\usage{
predGP_out(gpi, X, model.type = "laGP")
}
\arguments{
\item{model.type}{}
}
\description{
Make predictions from a Gaussian process, under different packages
}
