% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SKO.R
\name{SKO}
\alias{SKO}
\title{SKO algorithm}
\usage{
SKO(
  Xinit,
  f,
  N = 100,
  noisy = TRUE,
  sigma = 0,
  model.type = c("laGP", "km"),
  type = "UK",
  lower = 0,
  upper = 1,
  B = 2,
  beta = 0.9,
  predf = laGP::predGPsep,
  optim.multi.start = 5
)
}
\arguments{
\item{optim.multi.start}{}
}
\description{
Sequential Kriging Optimisation algorithm uses augmented expected improvement (AEI) with Gaussian processes
to find the minimum of a blackbox function
}
