% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSSO.R
\name{TSSO}
\alias{TSSO}
\title{TSSO algorithm}
\usage{
TSSO(
  Xinit,
  f,
  N = 100,
  noisy = TRUE,
  sigma = 0,
  model.type = c("laGP", "km"),
  type = "UK",
  lower = 0,
  upper = 1,
  B = 10,
  rmin = 2,
  predf = laGP::predGPsep,
  optim.multi.start = 5
)
}
\arguments{
\item{optim.multi.start}{}
}
\description{
Two-stage sequential optimisation (TSSO) algorithm uses modified expected improvement (MEI) to find
the minimum of a noisy blackbox function
}
